update stopwatch and wait()

+++ pybricks/tools/__stub/__tools.py
@@ -1,8 +1,8 @@
-def wait(time: float):
+def wait(time: int):
     """
     Pauses the user program for a specified amount of time.
 
     Args:
-        time (float): How long to wait in milliseconds.
+        time (int): How long to wait in milliseconds.
     """
     ...
